Fix warnings between floats and uint64

+++ cleanup.c
@@ -177,8 +177,8 @@ cleanup_dir(struct conf *conf, const char *dir)
 	// When "max files" or "max cache size" is reached, one of the 16 cache
 	// subdirectories is cleaned up. When doing so, files are deleted (in LRU
 	// order) until the levels are below limit_multiple.
-	cache_size_threshold = conf->max_size * conf->limit_multiple / 16;
-	files_in_cache_threshold = conf->max_files * conf->limit_multiple / 16;
+	cache_size_threshold = (uint64_t)(conf->max_size * conf->limit_multiple / 16);
+	files_in_cache_threshold = (uint64_t)(conf->max_files * conf->limit_multiple / 16);
 
 	num_files = 0;
 	cache_size = 0;
@@ -248,7 +248,7 @@ wipe_dir(struct conf *conf, const char *dir)
 {
 	cc_log("Clearing out cache directory %s", dir);
 
-	files_in_cache_threshold = conf->max_files * conf->limit_multiple / 16;
+	files_in_cache_threshold = (size_t)(conf->max_files * conf->limit_multiple / 16);
 	files_in_cache = 0;
 
 	traverse(dir, wipe_fn);
